// src/lib/types.ts
export type RequirementType = {
    id: string;
    complianceList: string;
    department: string;
    dateSubmitted: string;
    documentReference: string;
    entity: string;
    expiration: string;
    frequencyOfCompliance: string;
    personInCharge: string;
    renewal: string;
    status: string;
    typeOfCompliance: string;
  };
  